moves to the correct package for presenter and adapter
