--- app/src/main/java/com/affan/movieapp/view/main/movies/MoviesPresenter.kt
@@ -1,5 +1,0 @@
-package com.affan.movieapp.view.main.movies
-
-interface MoviesPresenter {
-    fun getMovies()
-}
