--- app/src/main/java/com/affan/movieapp/view/main/movies/MoviesView.kt
@@ -1,5 +1,0 @@
-package com.affan.movieapp.view.main.movies
-
-interface MoviesView {
-    fun onReceiveMovies(movies: List<MoviesData>)
-}
