+++ app/src/main/java/com/affan/movieapp/view/main/movies/presenter/MoviesPresenter.kt
@@ -1,0 +1,5 @@
+package com.affan.movieapp.view.main.movies.presenter
+
+interface MoviesPresenter {
+    fun getMovies()
+}
