+++ app/src/main/java/com/affan/movieapp/view/main/movies/presenter/MoviesView.kt
@@ -1,0 +1,7 @@
+package com.affan.movieapp.view.main.movies.presenter
+
+import com.affan.movieapp.view.main.movies.MoviesData
+
+interface MoviesView {
+    fun onReceiveMovies(movies: List<MoviesData>)
+}
